cxx,test: integrate the existing gccxml tests in the common C++ tests
