+++ test/ruby/cxx_import_tests/documentation_with_space_between_struct_and_opening_bracket.h
