+++ test/ruby/cxx_import_tests/documentation_with_struct_and_opening_bracket_on_different_lines.h
