+++ test/ruby/cxx_import_tests/ignored_base_class.h
