+++ test/ruby/cxx_import_tests/non_virtual_inheritance.h
