+++ test/ruby/cxx_import_tests/private_base_class.h
