+++ test/ruby/cxx_import_tests/template_of_container.h
