+++ test/ruby/cxx_import_tests/virtual_inheritance.h
