+++ test/ruby/suite_cxx.rb
@@ -1,3 +1,4 @@
+require_relative './test_cxx'
 require_relative './test_cxx_clang'
 require_relative './test_cxx_castxml'
 require_relative './test_cxx_gccxml'
