+++ test/ruby/test_cxx_clang.rb
@@ -1,11 +1,13 @@
 require 'typelib/test'
-require_relative './cxx_test_helpers'
+require_relative './cxx_common_tests'
 
 class TC_CXX_Clang < Minitest::Test
     include Typelib
+    include CXXCommonTests
 
-    extend CXXTestHelpers
-    cxx_test_dir = File.expand_path('cxx_import_tests', File.dirname(__FILE__))
-    generate_common_tests('clang', cxx_test_dir)
+    def setup
+        super
+        @loader = Typelib::CXX::CXX_LOADERS['clang']
+    end
 end
